Add VK ID authorization
1) Use VK ID for user authorization
2) Remove previous Django registration and
authorization forms.def login and def register
removed. register.html and login.html removed.
3) Improve page design
4) Change unit-tests for new authorization. Added djangi-tests.yml
for github Actions

Now the user on the main page sees the One Tap button for
authorization using VK ID. After a successful login, user
first receives a Silent token, and then an access token,
along with a redirect to the profile page.

+++ places_remember/main/views.py
@@ -1,53 +1,50 @@
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
-from django.contrib.auth import login as auth_login
 from django.shortcuts import render, redirect, get_object_or_404
-from django.contrib.auth.decorators import login_required
 from django.http import HttpResponseForbidden
 from .forms import MemoryForm
-from .models import Memory
-
-
+from .models import Memory, User
+from django.views.decorators.csrf import csrf_exempt
+from django.contrib.auth import logout
+from django.http import JsonResponse
+from django.shortcuts import render
+from django.conf import settings
+import requests
+import json
 # Create your views here.
+
 
 def home(request):
     return render(request, 'home.html')
 
 
-def login(request):
-    if request.method == 'POST':
-        form = AuthenticationForm(request, data=request.POST)
-        if form.is_valid():
-            auth_login(request, form.get_user())
-            return redirect('profile')
-    else:
-        form = AuthenticationForm()
-    return render(request, 'login.html', {'form': form})
+def logout_view(request):
+
+    request.session.flush()
+    logout(request)
+    return redirect('/')
 
 
-def register(request):
-    if request.method == 'POST':
-        form = UserCreationForm(request.POST)
-        if form.is_valid():
-            form.save()
-            return redirect('login')
-    else:
-        form = UserCreationForm()
-    return render(request, 'register.html', {'form': form})
-
-
-@login_required
 def profile(request):
-    memories = Memory.objects.filter(user=request.user)
-    return render(request, 'profile.html', {'memories': memories})
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('/')  
+
+    user = get_object_or_404(User, pk=user_id)
+    memories = user.memories.all()
+    return render(request, 'profile.html', {'user': user, 'memories': memories})
 
 
-@login_required
 def add_memory(request):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('/')
+
+    user = get_object_or_404(User, pk=user_id)
+
     if request.method == 'POST':
         form = MemoryForm(request.POST)
         if form.is_valid():
             memory = form.save(commit=False)
-            memory.user = request.user
+            memory.user = user
             latitude = request.POST.get('latitude')
             longitude = request.POST.get('longitude')
             if latitude is not None and longitude is not None:
@@ -60,20 +57,28 @@ def add_memory(request):
     return render(request, 'add_memory.html', {'form': form})
 
 
-@login_required
 def edit_memory(request, memory_id):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('/')
+
+    user = get_object_or_404(User, pk=user_id)
     memory = get_object_or_404(Memory, id=memory_id)
 
     # Проверка, является ли текущий пользователь владельцем воспоминания
-    if memory.user != request.user:
+    if memory.user != user:
         return HttpResponseForbidden("You are not allowed to edit this memory.")
 
     if request.method == 'POST':
         form = MemoryForm(request.POST, instance=memory)
         if form.is_valid():
-            # Сохранение значений формы в объект памяти
             memory = form.save(commit=False)
-            memory.user = request.user  # Обновляем пользователя, на всякий случай
+            memory.user = user
+
+            # Обновляем координаты
+            memory.latitude = form.cleaned_data['latitude']
+            memory.longitude = form.cleaned_data['longitude']
+
             memory.save()
             return redirect('profile')
     else:
@@ -83,8 +88,87 @@ def edit_memory(request, memory_id):
 
 
 def delete_memory(request, memory_id):
+    user_id = request.session.get('user_id')
+    if not user_id:
+        return redirect('/login/')
+
+    user = get_object_or_404(User, pk=user_id)
     memory = get_object_or_404(Memory, id=memory_id)
+
+    # Проверка, является ли текущий пользователь владельцем воспоминания
+    if memory.user != user:
+        return HttpResponseForbidden("You are not allowed to delete this memory.")
+
     if request.method == 'POST':
         memory.delete()
         return redirect('profile')
     return redirect('profile')
+
+
+@csrf_exempt
+def exchange_silent_token(request):
+    if request.method == 'POST':
+        data = json.loads(request.body)
+        silent_token = data.get('silent_token')
+        uuid = data.get('uuid')
+        service_token = settings.VK_SERVICE_TOKEN
+
+        # Запрос к VK API для обмена silent_token на access_token
+        response = requests.post('https://api.vk.com/method/auth.exchangeSilentAuthToken', data={
+            'v': '5.131',
+            'token': silent_token,
+            'access_token': service_token,
+            'uuid': uuid,
+        })
+
+        response_data = response.json()
+        if 'response' in response_data:
+            access_token = response_data['response']['access_token']
+            user_id = response_data['response']['user_id']
+
+            # Запрос к VK API для получения информации о пользователе
+            profile_response = requests.get('https://api.vk.com/method/account.getProfileInfo', params={
+                'v': '5.131',
+                'access_token': access_token
+            })
+
+            profile_data = profile_response.json()
+
+            if 'response' in profile_data:
+                user_info = profile_data['response']
+                avatar = data.get('user', {}).get('avatar')
+
+                if not avatar:
+                    # Если аватара нет в silent_token, сделаем запрос к users.get для получения фото
+                    user_response = requests.get('https://api.vk.com/method/users.get', params={
+                        'user_ids': user_id,
+                        'fields': 'photo_200',
+                        'v': '5.131',
+                        'access_token': access_token
+                    })
+
+                    user_data = user_response.json()
+                    if 'response' in user_data:
+                        avatar = user_data['response'][0].get('photo_200')
+
+                user_info['avatar'] = avatar
+
+                # Сохраните или обновите информацию о пользователе в базе данных
+                user, created = User.objects.update_or_create(
+                    vk_id=user_id,
+                    defaults={
+                        'first_name': user_info['first_name'],
+                        'last_name': user_info['last_name'],
+                        'avatar': avatar,
+                        'phone': user_info.get('phone')
+                    }
+                )
+
+                request.session['user_id'] = user.id
+                return JsonResponse({'access_token': access_token, 'user_info': user_info})
+            else:
+                return JsonResponse({'error': profile_data.get('error')}, status=400)
+        else:
+            return JsonResponse({'error': response_data.get('error')}, status=400)
+    return JsonResponse({'error': 'Invalid request method'}, status=405)
+
